pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "bug-triage-env"
version = "0.1.0"
description = "Bug Triage & Patch Validation OpenEnv Environment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
inference = [
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["bug_triage_env*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
